Рефакторит comments.js
